+++ src/arduino/boardManager.ts
@@ -13,9 +13,10 @@ import * as util from "../common/util";
 import { arduinoChannel } from "../common/outputChannel";
 import { DeviceContext } from "../deviceContext";
 import { ArduinoApp } from "./arduino";
+import { IArduinoSettings } from "./arduinoSettings";
 import { Board, parseBoardDescriptor } from "./board";
 import { IBoard, IPackage, IPlatform } from "./package";
-import { IArduinoSettings } from "./settings";
+import { VscodeSettings } from "./vscodeSettings";
 
 export class BoardManager {
 
@@ -49,13 +50,14 @@ export class BoardManager {
         this._platforms = [];
         this._installedPlatforms = [];
 
+        const addiontionalUrls = this.getAdditionalUrls();
         if (update) { // Update index files.
-            await this._arduinoApp.setPref("boardsmanager.additional.urls", this.getAdditionalUrls().join(","));
+            await this.setPreferenceUrls(addiontionalUrls);
             await this._arduinoApp.initialize(true);
         }
 
         // Parse package index files.
-        const indexFiles = ["package_index.json"].concat(this.getAdditionalUrls());
+        const indexFiles = ["package_index.json"].concat(addiontionalUrls);
         const rootPackgeFolder = this._settings.packagePath;
         for (const indexFile of indexFiles) {
             const indexFileName = this.getIndexFileName(indexFile);
@@ -63,7 +65,7 @@ export class BoardManager {
                 continue;
             }
             if (!update && !util.fileExistsSync(path.join(rootPackgeFolder, indexFileName))) {
-                await this._arduinoApp.setPref("boardsmanager.additional.urls", this.getAdditionalUrls().join(","));
+                await this.setPreferenceUrls(addiontionalUrls);
                 await this._arduinoApp.initialize(true);
             }
             this.loadPackageContent(indexFileName);
@@ -76,6 +78,11 @@ export class BoardManager {
         this.loadInstalledBoards();
         this.updateStatusBar();
         this._boardStatusBar.show();
+
+        const dc = DeviceContext.getIntance();
+        dc.onDidChange(() => {
+            this.updateStatusBar();
+        });
     }
 
     public async changeBoardType() {
@@ -105,13 +112,11 @@ export class BoardManager {
 
     public async updatePackageIndex(indexUri: string): Promise<boolean> {
         const indexFileName = this.getIndexFileName(indexUri);
-        if (util.fileExistsSync(path.join(this._settings.packagePath, indexFileName))) {
-            return false;
-        }
+
         let allUrls = this.getAdditionalUrls();
         if (!(allUrls.indexOf(indexUri) >= 0)) {
             allUrls = allUrls.concat(indexUri);
-            await this._settings.updateAdditionalUrls(allUrls);
+            await VscodeSettings.getIntance().updateAdditionalUrls(allUrls);
             await this._arduinoApp.setPref("boardsmanager.additional.urls", this.getAdditionalUrls().join(","));
         }
         return true;
@@ -150,8 +155,8 @@ export class BoardManager {
     public getInstalledPlatforms(): any[] {
         // Always using manually installed platforms to overwrite the same platform from arduino installation directory.
         const installedPlatforms = this.getDefaultPlatforms();
-        const manuallyInstalled = this.getManuallyInstalledPlatforms();
-        manuallyInstalled.forEach((plat) => {
+
+        const mergePlatform = (plat) => {
             const find = installedPlatforms.find((_plat) => {
                 return _plat.packageName === plat.packageName && _plat.architecture === plat.architecture;
             });
@@ -162,7 +167,14 @@ export class BoardManager {
                 find.version = plat.version;
                 find.rootBoardPath = plat.rootBoardPath;
             }
-        });
+        };
+
+        const customPlatforms = this.getCustomPlatforms();
+        const manuallyInstalled = this.getManuallyInstalledPlatforms();
+
+        customPlatforms.forEach(mergePlatform);
+        manuallyInstalled.forEach(mergePlatform);
+
         return installedPlatforms;
     }
 
@@ -182,7 +194,7 @@ export class BoardManager {
         } catch (ex) {
             arduinoChannel.error(`Invalid json file "${path.join(this._settings.packagePath, indexFileName)}".
             Suggest to remove it manually and allow boardmanager to re-download it.`);
-            return ;
+            return;
         }
 
         this._packages.concat(rawModel.packages);
@@ -273,12 +285,15 @@ export class BoardManager {
                     existingPlatform.rootBoardPath = platform.rootBoardPath;
                     this._installedPlatforms.push(existingPlatform);
                 }
+            } else {
+                platform.installedVersion = platform.version;
+                this._installedPlatforms.push(platform);
             }
         });
     }
 
     // Default arduino package information from arduino installation directory.
-    private getDefaultPlatforms(): any[] {
+    private getDefaultPlatforms(): IPlatform[] {
         const defaultPlatforms = [];
         try {
             const packageBundled = fs.readFileSync(path.join(this._settings.defaultPackagePath, "package_index_bundled.json"), "utf8");
@@ -304,6 +319,39 @@ export class BoardManager {
         } catch (ex) {
         }
         return defaultPlatforms;
+    }
+
+    private getCustomPlatforms(): IPlatform[] {
+        const customPlatforms = [];
+        const hardwareFolder = path.join(this._settings.sketchbookPath, "hardware");
+        if (!util.directoryExistsSync(hardwareFolder)) {
+            return customPlatforms;
+        }
+
+        const dirs = util.filterJunk(util.readdirSync(hardwareFolder, true)); // in Mac, filter .DS_Store file.
+        if (!dirs || dirs.length < 1) {
+            return customPlatforms;
+        }
+        for (const packageName of dirs) {
+            const architectures = util.filterJunk(util.readdirSync(path.join(hardwareFolder, packageName), true));
+            if (!architectures || architectures.length < 1) {
+                continue;
+            }
+            architectures.forEach((architecture) => {
+                const platformFolder = path.join(hardwareFolder, packageName, architecture);
+                if (util.fileExistsSync(path.join(platformFolder, "boards.txt")) && util.fileExistsSync(path.join(platformFolder, "platform.txt"))) {
+                    const configs = util.parseConfigFile(path.join(platformFolder, "platform.txt"));
+                    customPlatforms.push({
+                        packageName,
+                        architecture,
+                        version: configs.get("version"),
+                        rootBoardPath: path.join(hardwareFolder, packageName, architecture),
+                        defaultPlatform: false,
+                    });
+                }
+            });
+        }
+        return customPlatforms;
     }
 
     // User manually installed packages.
@@ -384,12 +432,19 @@ export class BoardManager {
             return [];
         }
         // For better compatibility, merge urls both in user settings and arduino IDE preferences.
-        const settingsUrls = formatUrls(this._settings.additionalUrls);
+        const settingsUrls = formatUrls(VscodeSettings.getIntance().additionalUrls);
         let preferencesUrls = [];
-        const preferences = this._arduinoApp.preferences;
+        const preferences = this._settings.preferences;
         if (preferences && preferences.has("boardsmanager.additional.urls")) {
             preferencesUrls = formatUrls(preferences.get("boardsmanager.additional.urls"));
         }
         return util.union(settingsUrls, preferencesUrls);
+    }
+
+    private async setPreferenceUrls(addiontionalUrls: string[]) {
+        const settingsUrls = addiontionalUrls.join(",");
+        if (this._settings.preferences.get("boardsmanager.additional.urls") !== settingsUrls) {
+            await this._arduinoApp.setPref("boardsmanager.additional.urls", settingsUrls);
+        }
     }
 }