+++ src/common/util.ts
@@ -9,6 +9,7 @@ import * as os from "os";
 import * as path from "path";
 import * as properties from "properties";
 import * as vscode from "vscode";
+import * as WinReg from "winreg";
 
 /**
  * This function will return the VSCode C/C++ extesnion compatible platform literals.
@@ -309,4 +310,58 @@ export function padStart(sourceString: string, targetLength: number, padString?:
     } else {
         return sourceString.padStart(targetLength, padString);
     }
+}
+
+export function parseConfigFile(fullFileName, filterComment: boolean = true): Map<string, string> {
+    const result = new Map<string, string>();
+    const lineRegex = /(\S+)=(\S+)/;
+
+    if (fileExistsSync(fullFileName)) {
+        const rawText = fs.readFileSync(fullFileName, "utf8");
+        const lines = rawText.split("\n");
+        lines.forEach((line) => {
+            if (line) {
+                if (filterComment) {
+                    if (line.trim() && line.startsWith("#")) {
+                        return;
+                    }
+                }
+                const match = lineRegex.exec(line);
+                if (match && match.length > 2) {
+                    result.set(match[1], match[2]);
+                }
+            }
+        });
+    }
+    return result;
+}
+
+export function getRegistryValues(hive: string, key: string, name: string): Promise<string> {
+    return new Promise((resolve, reject) => {
+        try {
+            const regKey = new WinReg({
+                hive,
+                key,
+            });
+
+            regKey.valueExists(name, (e, exists) => {
+                if (e) {
+                    return reject(e);
+                }
+                if (exists) {
+                    regKey.get(name, (err, result) => {
+                        if (!err) {
+                            resolve(result ? result.value : "");
+                        } else {
+                            reject(err);
+                        }
+                    });
+                } else {
+                    resolve("");
+                }
+            });
+        } catch (ex) {
+            reject(ex);
+        }
+    });
 }