+++ test/extension.test.ts
@@ -44,6 +44,7 @@ suite("Arduino Extension Tests", () => {
                     "arduino.changeBaudRate",
                     "arduino.sendMessageToSerialPort",
                     "arduino.closeSerialMonitor",
+                    "arduino.reloadExample",
                 ];
 
                 const foundArduinoCommands = commands.filter((value) => {
@@ -59,7 +60,7 @@ suite("Arduino Extension Tests", () => {
         // When running test on osx, the vscode instance is hanging there after tests finished and cause mocha timeout.
         // As a workaround, closing usb-detection process manually would make test window exit normally.
         if (os.platform() !== "linux") {
-            const usbDector = require("../../vendor/node-usb-detection");
+            const usbDector = require("../../vendor/node-usb-native").detector;
             usbDector.stopMonitoring();
         }
     });